Migrate to new command handling system
